bug: Updated SettingsScreen to reroute user back to loginscreen if they try to go back to the setting screeen after loggin out. * Updated gamewindow with new updatescore method that only stores a highscore if it is higher than the user's existing highscore. * Updated settings screen with new login screen button and removed sound slides as the user can just use their hardware volumen switch. Updated scores screen to display the user's highscore. * Updated help menu to display video content from the server. This needs to be tested on a physical device as its not displaying the video content in the emulator but shows the video as playing. * Uploaded demo.mp4 to the server. * bug: fixed deleteUser to remove user from both remote and local databases but not add to the local database after trying to delete the same user twice in a row.

+++ app/src/main/java/storage/shapes/LocalDatabaseOperations.java
@@ -18,7 +18,6 @@ import java.util.ArrayList;
  * Created by Fritz on 4/6/2016.
  */
 class LocalDatabaseOperations implements LocalDbOperations, SharedDbOperations  {
-    private Context context;
     private LocalDbHandler localDbHandler;
     private SQLiteDatabase sqLiteDatabase;
     private String sqlCmd;
@@ -26,7 +25,6 @@ class LocalDatabaseOperations implements LocalDbOperations, SharedDbOperations  
 
     /* Begin constructors */
     LocalDatabaseOperations(Context context) {
-        this.context = context;
         localDbHandler = new LocalDbHandler(context, "shapes", 1);
         sqLiteDatabase = localDbHandler.getWritableDatabase();
     }
@@ -224,26 +222,34 @@ class LocalDatabaseOperations implements LocalDbOperations, SharedDbOperations  
 
     public boolean deleteUser(String username, String password)
     {
-        sqlCmd = "delete from user " +
-                "where username = '" + username + "'";
+        String col[] = {"username"};
+        String selArgs[] = {username};
 
-        try {
-            sqLiteDatabase.execSQL(sqlCmd);
-        } catch (SQLiteConstraintException e) {
-            ;
-        }
+        cursor = sqLiteDatabase.query("user", col, "username = ?", selArgs, null, null, null);
 
-        sqlCmd = "delete from friends " +
-                "where user = '" + username + "'";
+        if (cursor.getCount() <= 0) {
+            return false;
+        } else {
+            sqlCmd = "delete from user " +
+                    "where username = '" + username + "'";
 
-        try {
-            sqLiteDatabase.execSQL(sqlCmd);
+            try {
+                sqLiteDatabase.execSQL(sqlCmd);
+            } catch (SQLiteConstraintException e) {
+                ;
+            }
+
+            sqlCmd = "delete from friends " +
+                    "where user = '" + username + "'";
+
+            try {
+                sqLiteDatabase.execSQL(sqlCmd);
+            } catch (SQLiteConstraintException e) {
+                ;
+            }
+
             return true;
-        } catch (SQLiteConstraintException e) {
-            ;
         }
-
-        return false;
     }
 
     public boolean setPassword(String username, String oldPassword, String newPassword)
